Move API lookup from BeatmapDetailArea to Leaderboard.
